requirements.txt
playwright>=1.40.0
pypdf>=3.0.0
pytest>=7.0.0
pytest-asyncio>=0.24.0
Pillow>=10.0.0
imagehash>=4.3.0